perf: 归档任务情况增加监控指标 #3824
处理review意见
