+++ src/backend/job-backup/service-job-backup/src/main/java/com/tencent/bk/job/backup/archive/dao/impl/GseFileAgentTaskRecordDAO.java
@@ -33,7 +33,6 @@ import org.jooq.Condition;
 import org.jooq.OrderField;
 import org.jooq.TableField;
 
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -57,11 +56,6 @@ public class GseFileAgentTaskRecordDAO extends AbstractJobInstanceHotRecordDAO<G
     @Override
     public TableField<GseFileAgentTaskRecord, Long> getJobInstanceIdField() {
         return TABLE.TASK_INSTANCE_ID;
-    }
-
-    @Override
-    protected TableField<GseFileAgentTaskRecord, LocalDateTime> getTableCreateTimeField() {
-        return TABLE.ROW_CREATE_TIME;
     }
 
     @Override
