+++ src/backend/job-backup/service-job-backup/src/main/java/com/tencent/bk/job/backup/archive/dao/impl/JobInstanceHotRecordDAO.java
@@ -30,7 +30,6 @@ import com.tencent.bk.job.execute.model.tables.records.TaskInstanceRecord;
 import org.jooq.OrderField;
 import org.jooq.TableField;
 
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -55,11 +54,6 @@ public class JobInstanceHotRecordDAO extends AbstractJobInstanceMainHotRecordDAO
     @Override
     public TableField<TaskInstanceRecord, Long> getJobInstanceIdField() {
         return TABLE.ID;
-    }
-
-    @Override
-    protected TableField<TaskInstanceRecord, LocalDateTime> getTableCreateTimeField() {
-        return TABLE.ROW_CREATE_TIME;
     }
 
     @Override
