+++ src/backend/job-backup/service-job-backup/src/main/java/com/tencent/bk/job/backup/archive/dao/impl/RollingConfigRecordDAO.java
@@ -32,7 +32,6 @@ import com.tencent.bk.job.execute.model.tables.records.RollingConfigRecord;
 import org.jooq.OrderField;
 import org.jooq.TableField;
 
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -62,11 +61,6 @@ public class RollingConfigRecordDAO extends AbstractJobInstanceHotRecordDAO<Roll
     @Override
     public TableField<RollingConfigRecord, Long> getJobInstanceIdField() {
         return TABLE.TASK_INSTANCE_ID;
-    }
-
-    @Override
-    protected TableField<RollingConfigRecord, LocalDateTime> getTableCreateTimeField() {
-        return TABLE.ROW_CREATE_TIME;
     }
 
     public RecordResultSet<RollingConfigRecord> executeQuery(Collection<Long> jobInstanceIds,
