+++ src/backend/job-backup/service-job-backup/src/main/java/com/tencent/bk/job/backup/archive/dao/impl/StepInstanceFileRecordDAO.java
@@ -33,7 +33,6 @@ import org.jooq.Condition;
 import org.jooq.OrderField;
 import org.jooq.TableField;
 
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -60,11 +59,6 @@ public class StepInstanceFileRecordDAO extends AbstractJobInstanceHotRecordDAO<S
     @Override
     public TableField<StepInstanceFileRecord, Long> getJobInstanceIdField() {
         return TABLE.TASK_INSTANCE_ID;
-    }
-
-    @Override
-    protected TableField<StepInstanceFileRecord, LocalDateTime> getTableCreateTimeField() {
-        return TABLE.ROW_CREATE_TIME;
     }
 
     protected Collection<? extends OrderField<?>> getListRecordsOrderFields() {
