+++ src/backend/job-backup/service-job-backup/src/main/java/com/tencent/bk/job/backup/archive/dao/impl/StepInstanceVariableRecordDAO.java
@@ -33,7 +33,6 @@ import org.jooq.Condition;
 import org.jooq.OrderField;
 import org.jooq.TableField;
 
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -58,11 +57,6 @@ public class StepInstanceVariableRecordDAO extends AbstractJobInstanceHotRecordD
     @Override
     public TableField<StepInstanceVariableRecord, Long> getJobInstanceIdField() {
         return TABLE.TASK_INSTANCE_ID;
-    }
-
-    @Override
-    protected TableField<StepInstanceVariableRecord, LocalDateTime> getTableCreateTimeField() {
-        return TABLE.ROW_CREATE_TIME;
     }
 
     protected Collection<? extends OrderField<?>> getListRecordsOrderFields() {
