+++ src/backend/job-backup/service-job-backup/src/main/java/com/tencent/bk/job/backup/archive/dao/impl/TaskInstanceHostRecordDAO.java
@@ -33,7 +33,6 @@ import org.jooq.Condition;
 import org.jooq.OrderField;
 import org.jooq.TableField;
 
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -56,11 +55,6 @@ public class TaskInstanceHostRecordDAO extends AbstractJobInstanceHotRecordDAO<T
     @Override
     public TableField<TaskInstanceHostRecord, Long> getJobInstanceIdField() {
         return TABLE.TASK_INSTANCE_ID;
-    }
-
-    @Override
-    protected TableField<TaskInstanceHostRecord, LocalDateTime> getTableCreateTimeField() {
-        return TABLE.ROW_CREATE_TIME;
     }
 
     protected Collection<? extends OrderField<?>> getListRecordsOrderFields() {
