--- src/backend/job-backup/service-job-backup/src/main/java/com/tencent/bk/job/backup/archive/service/detector/UnarchivedRollingConfigDetector.java
@@ -1,47 +1,0 @@
-/*
- * Tencent is pleased to support the open source community by making BK-JOB蓝鲸智云作业平台 available.
- *
- * Copyright (C) 2021 Tencent.  All rights reserved.
- *
- * BK-JOB蓝鲸智云作业平台 is licensed under the MIT License.
- *
- * License for BK-JOB蓝鲸智云作业平台:
- * --------------------------------------------------------------------
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
- * documentation files (the "Software"), to deal in the Software without restriction, including without limitation
- * the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and
- * to permit persons to whom the Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of
- * the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO
- * THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF
- * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
- * IN THE SOFTWARE.
- */
-
-package com.tencent.bk.job.backup.archive.service.detector;
-
-import com.tencent.bk.job.backup.archive.dao.ArchiveTaskDAO;
-import com.tencent.bk.job.backup.archive.dao.impl.RollingConfigRecordDAO;
-import com.tencent.bk.job.backup.config.ArchiveProperties;
-import io.micrometer.core.instrument.MeterRegistry;
-
-public class UnarchivedRollingConfigDetector extends UnarchivedDataDetector {
-
-    private final String tableName = "job_execute.rolling_config";
-
-    public UnarchivedRollingConfigDetector(MeterRegistry meterRegistry,
-                                           ArchiveTaskDAO archiveTaskDAO,
-                                           RollingConfigRecordDAO rollingConfigRecordDAO,
-                                           ArchiveProperties archiveProperties) {
-        super(meterRegistry, archiveTaskDAO, archiveProperties, rollingConfigRecordDAO);
-    }
-
-    @Override
-    protected String getTableName() {
-        return tableName;
-    }
-}
